# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: VJU Club Management API
  version: 1.0.0
  description: 'API giai đoạn 1 cho quản lý tài khoản, CLB, ban, thành viên và permission.
    Xem x-implementation-status từng operation: IMPLEMENTED là đã xác nhận qua integration
    test. UUID, camelCase, thời gian ISO-8601. List dùng offset/limit; không có page/size.'
servers:
- url: /api/v1
  description: Backend cùng origin với Swagger UI
- url: http://localhost:8080/api/v1
  description: Backend local mặc định
security:
- sessionAuth: []
paths:
...
  /auth/login:
    post:
      operationId: auth.login
      summary: Login
      tags:
      - Auth
      x-permission: null
      x-implementation-status: IMPLEMENTED
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
              example:
                user:
                  id: 11111111-1111-4111-8111-111111111111
                  email: phat@example.test
                  fullName: Nguyễn Xuân Phát
                  studentCode: '23110209'
                  phone: 0900000000
                  avatarUrl: null
                  status: ACTIVE
                  createdAt: '2026-09-29T12:00:00Z'
                  updatedAt: '2026-09-29T12:00:00Z'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '400':
          $ref: '#/components/responses/BadRequest'
        '409':
          $ref: '#/components/responses/Conflict'
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
            example:
              email: phat@example.test
              password: DemoPassword123!
      description: 'Trả user và đặt cookie phiên CLUB_SESSION (HttpOnly, SameSite=Lax). Không có token trong body.
        Gửi kèm header X-XSRF-TOKEN lấy từ cookie XSRF-TOKEN (GET /auth/csrf). Tài khoản INACTIVE bị từ chối (403).
        Tài khoản do admin cấp (POST /users); quên mật khẩu liên hệ admin.'
...
  /auth/me:
    get:
      operationId: auth.me
      summary: Get current account
      tags:
      - Auth
      x-permission: null
      x-implementation-status: IMPLEMENTED
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
              example:
                id: 11111111-1111-4111-8111-111111111111
                email: phat@example.test
                fullName: Nguyễn Xuân Phát
                studentCode: '23110209'
                phone: 0900000000
                avatarUrl: null
                status: ACTIVE
                createdAt: '2026-09-29T12:00:00Z'
                updatedAt: '2026-09-29T12:00:00Z'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '400':
          $ref: '#/components/responses/BadRequest'
      description: Tài khoản hiện tại theo JWT. Chỉ cần đăng nhập; không cần quyền
        quản trị user.view.
  /auth/change-password:
    post:
      operationId: auth.changePassword
      summary: Change password
      tags:
      - Auth
      x-permission: null
      x-implementation-status: IMPLEMENTED
      responses:
        '204':
          description: Password changed
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '409':
          $ref: '#/components/responses/Conflict'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChangePasswordRequest'
            example:
              currentPassword: DemoPassword123!
              newPassword: NewDemoPassword456!
      description: Xác minh currentPassword, đổi mật khẩu và thu hồi refresh token.
        Sau 204, frontend xóa token và yêu cầu đăng nhập lại.
...
components:
  securitySchemes:
    sessionAuth:
      type: apiKey
      in: cookie
      name: CLUB_SESSION
      description: 'Cookie phiên do POST /auth/login đặt. Request thay đổi dữ liệu cần thêm header X-XSRF-TOKEN.'
  responses:
    BadRequest:
      description: Invalid request
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
    Unauthorized:
      description: 'Chưa đăng nhập hoặc phiên đã hết hạn/bị huỷ (code UNAUTHORIZED); đăng nhập lại.'
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
    Forbidden:
      description: Permission denied
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
    Conflict:
      description: Trùng dữ liệu hoặc trạng thái hiện tại không cho phép thao tác.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
...
  schemas:
...
    LoginRequest:
      type: object
      required:
      - email
      - password
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          format: password
      examples:
      - email: phat@example.test
        password: DemoPassword123!
    ChangePasswordRequest:
      type: object
      required:
      - currentPassword
      - newPassword
      properties:
        currentPassword:
          type: string
          format: password
        newPassword:
          type: string
          format: password
          minLength: 8
          maxLength: 72
          description: '8+ ký tự, tối đa 72 byte UTF-8 (giới hạn của BCrypt; chữ có dấu tốn 2–3 byte)'
      examples:
      - currentPassword: DemoPassword123!
        newPassword: NewDemoPassword456!
    AuthResponse:
      type: object
      required:
      - user
      properties:
        user:
          $ref: '#/components/schemas/User'
    User:
      type: object
      properties:
        id:
          type: string
          format: uuid
          readOnly: true
        email:
          type: string
          format: email
        fullName:
          type: string
        studentCode:
          type:
          - string
          - 'null'
        phone:
          type:
          - string
          - 'null'
        avatarUrl:
          type:
          - string
          - 'null'
        status:
          type: string
          enum:
          - ACTIVE
          - INACTIVE
        createdAt:
          type: string
          format: date-time
          readOnly: true
        updatedAt:
          type: string
          format: date-time
          readOnly: true
      required:
      - id
      - email
      - fullName
      - studentCode
      - phone
      - avatarUrl
      - status
      - createdAt
      - updatedAt
      examples:
      - id: 11111111-1111-4111-8111-111111111111
        email: phat@example.test
        fullName: Nguyễn Xuân Phát
        studentCode: '23110209'
        phone: 0900000000
        avatarUrl: null
        status: ACTIVE
        createdAt: '2026-09-29T12:00:00Z'
        updatedAt: '2026-09-29T12:00:00Z'
...
    Problem:
      type: object
      properties:
        type:
          type: string
          format: uri-reference
        title:
          type: string
        status:
          type: integer
        detail:
          type: string
        instance:
          type: string
          format: uri-reference
          description: URI request, có thể vắng với lỗi do security filter trả về.
        code:
          type: string
      required:
      - type
      - title
      - status
      - detail
      - code
      examples:
      - type: about:blank
        title: PERMISSION_DENIED
        status: 403
        detail: Permission denied
        instance: /api/v1/clubs/22222222-2222-4222-8222-222222222222
        code: PERMISSION_DENIED
...